% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{black_scholes_example}
\alias{black_scholes_example}
\title{Black-Scholes via AD}
\usage{
black_scholes_example()
}
\value{
No value is returned as the function prints to stdout
}
\description{
This example illustrate how to use automatic differentiation to
calculate the delte of a Black-Scholes call and put. It is based
on the same example in the FastAD sources.
}
\examples{
black_scholes_example()
}
