% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{linear_regression}
\alias{linear_regression}
\title{Evaluate a squared-loss linear regression at a given parameter value}
\usage{
linear_regression(X, y, theta_hat, initial_lr = 0.1, max_iter = 100L,
  tol = 1e-07)
}
\arguments{
\item{X}{Matrix with independent explanatory variables}

\item{y}{Vector with dependent variable}

\item{theta_hat}{Vector of parameters}
}
\value{
A list object with the \sQuote{loss} and \sQuote{gradient}
}
\description{
Not that this function does not actually fit the model. Rather it evaluates
the squared sum of residuals and \sQuote{gradient} of parameters.
}
\examples{
data(trees)   # also used in help(lm)
X <- as.matrix(cbind(const=1, trees[, c("Girth", "Height")]))
y <- trees$Volume
linear_regression(X, y, rep(0, 3), tol=1e-12)
coef(lm(y ~ X - 1))  # for comparison
}
