% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{quadratic_expression}
\alias{quadratic_expression}
\title{Compute the value and derivate of a quadratic expression X' * Sigma * X}
\usage{
quadratic_expression(X, Sigma)
}
\arguments{
\item{X}{A 2 element vector}

\item{Sigma}{A 2 x 2 matrix}
}
\value{
A list with two elements for the expression evaluated for X and Sigma as well as
}
\description{
Compute the value and derivate of a quadratic expression X' * Sigma * X
}
\examples{
X <- c(0.5, 0.6)
S <- matrix(c(2, 3, 3, 6), 2, 2)
quadratic_expression(X, S)
}
