\name{RcppFastAD-package}
\alias{RcppFastAD-package}
\alias{RcppFastAD}
\docType{package}
\title{\packageTitle{RcppFastAD}}
\description{\packageDescription{RcppFastAD}}
\section{Package Content}{\packageIndices{RcppFastAD}}
\author{\packageAuthor{RcppFastAD}}
\section{Maintainer}{\packageMaintainer{RcppFastAD}}
\keyword{package}
