% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{black_scholes}
\alias{black_scholes}
\title{Black-Scholes via AD}
\usage{
black_scholes(spot = 105, strike = 100, vol = 5, r = 1.25/100,
  tau = 30/365)
}
\value{
A matrix with rows for the call and put variant, and columns
for option value, delta and vega
}
\description{
This example illustrate how to use automatic differentiation to
calculate the delte of a Black-Scholes call and put. It is based
on the same example in the FastAD sources.
}
\examples{
black_scholes_example()
}
