% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{linear_regression}
\alias{linear_regression}
\title{Evaluate a squared-loss linear regression at a given parameter value}
\usage{
linear_regression(theta_hat, y, X)
}
\arguments{
\item{theta_hat}{Vector of parameters}

\item{y}{Vector with dependent variable}

\item{X}{Matrix with independent explanatory variables}
}
\value{
A list object with the \sQuote{loss} and \sQuote{gradient}
}
\description{
Not that this function does not actually fit the model. Rather it evaluates
the squared sum of residuals and \sQuote{gradient} of parameters.
}
\examples{
beta_guess <- c(1,2)
y <- c(32,64,96,128,160)
X <- matrix(c(1,10,2,20,3,30,4,40,5,50), ncol=2, byrow=TRUE)
linear_regression(beta_guess, y, X)
}
